Bare bones authentication
Currently just running it using localhost on port 5000. Not using a proper running site.
Can go through the stage of authentication and grabbing the code manually and finally making a simple POST request.
I will want to clean this up to allow for a nicer flow where the user does not need to manually get the code